Ready to use for user tracking Mongo trend freq counter needs testing

+++ src/TweetAnalytics/DBManager.java
@@ -14,7 +14,10 @@ import java.util.Random;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import twitter4j.api.TweetsResources;
+
 import com.mongodb.AggregationOptions;
+import com.mongodb.BasicDBList;
 import com.mongodb.BasicDBObject;
 import com.mongodb.Cursor;
 import com.mongodb.DB;
@@ -58,6 +61,12 @@ public class DBManager {
 	 */
 	@SuppressWarnings("deprecation")
 	public void groupTweetsByUser() {
+
+		// In order to avoid deleting potentially large amounts of data..again
+		if (db.collectionExists("users"))
+			throw new RuntimeException(
+					"Mongo: Users collection already exists.");
+
 		db.getCollection("users").drop();
 		db.getCollection("users").ensureIndex(
 				new BasicDBObject("tweets", "text"));
@@ -100,79 +109,148 @@ public class DBManager {
 	}
 
 	@SuppressWarnings("deprecation")
-	public void countAppearanceByUser() {
+	/**
+	 * Counts the number of trending topics each user appears in and saves it as a
+	 * field in the user collection
+	 */
+	public void countFrequencyByUser() {
 
 		// Index user collection by freq count in desc order
-		db.getCollection("users").ensureIndex(
-				new BasicDBObject("frequency", -1));
+		db.getCollection("users")
+				.ensureIndex(new BasicDBObject("frequency", 1));
 
 		Cursor trends = db.getCollection("trends").find();
+		/*
+		 * while (trends.hasNext()) { // db.users.update({ // $text : { $search
+		 * : "\"<Trend>\"", $language : "none"} }, // { $inc : { frequency: 1 }
+		 * }, // { multi : true }); // Note: Escaping hyphens for phrase search
+		 * String s = (String) trends.next().get("name"); System.out.println(s);
+		 * DBObject query = new BasicDBObject("$text", new BasicDBObject(
+		 * "$search", "\"" + s + "\"").append("$language", "none"));
+		 * 
+		 * DBObject updateOp = new BasicDBObject("$inc", new BasicDBObject(
+		 * "frequency", 1));
+		 * 
+		 * db.getCollection("users").update(query, updateOp, false, true); }
+		 */
 
-		while (trends.hasNext()) {
-			// db.users.update({
-			// $text : { $search : "\"<Trend>\""} },
-			// { $inc : { frequency: 1 } },
-			// { multi : true });
-			// Note: Escaping hyphens for phrase search
-			String s = (String) trends.next().get("name");
-			System.out.println(s);
-			DBObject query = new BasicDBObject("$text", new BasicDBObject(
-					"$search", "\"" + s + "\""));
+		// Pure java version, most likely faster
+		List<String> topics = new ArrayList<>();
+		while (trends.hasNext())
+			topics.add((String) trends.next().get("name"));
+		trends.close();
 
-			DBObject updateOp = new BasicDBObject("$inc", new BasicDBObject(
-					"frequency", 1));
-
-			db.getCollection("users").update(query, updateOp, false, true);
+		Cursor users = db.getCollection("users").find();
+		int i = 0;
+		while (users.hasNext()) {
+			DBObject user = users.next();
+			System.out.println("User " + i++);
+			@SuppressWarnings("unchecked")
+			List<String> tweets = (List<String>) (List<?>) (BasicDBList) user
+					.get("tweets");
+			int freq = 0;
+			for (String tr : topics) {
+				for (String tw : tweets)
+					if (tw.toLowerCase().contains(tr.toLowerCase())) {
+						++freq;
+						break; // If you find a tweet matching the topic, 
+								// don't check the rest
+					}
+			}
+			// Update users appearance frequency in trends
+			db.getCollection("users").update(
+					new BasicDBObject("$match", new BasicDBObject("_id",
+							user.get("_id"))),
+					new BasicDBObject("$set", new BasicDBObject("frequency",
+							freq)), false, false);
 		}
+		users.close();
 	}
 
-	public int[] calculateQuartiles() {
+	@SuppressWarnings("deprecation")
+	public int[] calculateQuartiles(String collection) {
 
-		long c = db.getCollection("users").getCount();
+		long c = db.getCollection(collection).getCount();
 		System.out.println("Document count: " + c);
 		long first_halve = c / 4, second_halve = c / 2, third_halve = 3 * c / 4;
-		int q1 = 0, q2 = 0, q3 = 0, index = 0;
-
-		Cursor users = db.getCollection("users").find()
+		int q1, q2, q3, index = 0;
+		db.getCollection(collection).ensureIndex(
+				new BasicDBObject("frequency", 1));
+		Cursor users = db.getCollection(collection).find()
 				.sort(new BasicDBObject("frequency", 1));
-		// boolean even = (c % 2 == 0);
 
 		while (index++ < first_halve)
 			users.next();
 		System.out.println("ch1: " + index);
-		q1 = (int) users.next().get("frequency");
+		q1 = getFreq(users.next());
 
 		while (index++ < second_halve)
 			users.next();
 		System.out.println("ch2: " + index);
-		q2 = (int) users.next().get("frequency");
+		q2 = getFreq(users.next());
 
 		while (index++ < third_halve)
 			users.next();
 		System.out.println("ch3: " + index);
-		q3 = (int) users.next().get("frequency");
+		q3 = getFreq(users.next());
 
+		users.close();
 		return new int[] { q1, q2, q3 };
 	}
 
-	public long[] pickRandomUsersPerQuartile(int[] quartiles, int num_users) {
+	/**
+	 * If there is no frequency field (text doesn't belong to any trend(wonders
+	 * of twitter API)) return 0
+	 * 
+	 * @param usr
+	 * @return frequency value
+	 */
+	int getFreq(DBObject usr) {
+		if (usr.get("frequency") == null)
+			return 0;
+		return (int) usr.get("frequency");
+	}
+
+	public long[] pickUsersPerQuartile(String inCol, String outCol,
+			int[] quartiles, int num_users) {
+
 		long[] user_ids = new long[num_users * 4];
-		long c = db.getCollection("users").getCount();
-		int qr = (int) (0.25 * c);
-		Random n = new Random(Calendar.getInstance().getTimeInMillis());
-		for (int q = 0; q < 4; q++) {
-			for (int i = 0; i < num_users; i++) {
-				int l = (int) (q * qr) + n.nextInt(qr);
-				// System.out.println(q + " " + c + " " + l);
-				DBObject e = db.getCollection("users").find()
-						.sort(new BasicDBObject("frequency", 1)).limit(-1)
-						.skip(l).next();
-				e.put("quartile", q);
-				db.getCollection("chosen_users").insert(e);
+		// If collection exists, fetch users, else create it
+		if (db.collectionExists(outCol)) {
+			Cursor cursor = db.getCollection(outCol).find();
+			int i = 0;
+			while (cursor.hasNext()) {
+				DBObject e = cursor.next();
 				if (e.get("_id") instanceof Long)
-					user_ids[(int)(q * num_users + i)] = (Long) e.get("_id");
+					user_ids[i++] = (Long) e.get("_id");
 				else if (e.get("_id") instanceof Integer)
-					user_ids[(int)(q * num_users + i)] = ((Integer) e.get("_id")).longValue();
+					user_ids[i++] = ((Integer) e.get("_id")).longValue();
+			}
+		} else {
+
+			long c = db.getCollection(inCol).getCount();
+			int qr = (int) (0.25 * c);
+			Random n = new Random(Calendar.getInstance().getTimeInMillis());
+			for (int q = 0; q < 4; q++) {
+				for (int i = 0; i < num_users; i++) {
+					
+					// System.out.println(q + " " + c + " " + l);
+					DBObject e;
+					do {
+						int l = (int) (q * qr) + n.nextInt(qr);
+						e = db.getCollection(inCol).find()
+								.sort(new BasicDBObject("frequency", 1))
+								.limit(-1).skip(l).next();
+					} while (e.get("frequency") == null);
+					e.put("quartile", q);
+					db.getCollection(outCol).insert(e);
+					if (e.get("_id") instanceof Long)
+						user_ids[(int) (q * num_users + i)] = (Long) e
+								.get("_id");
+					else if (e.get("_id") instanceof Integer)
+						user_ids[(int) (q * num_users + i)] = ((Integer) e
+								.get("_id")).longValue();
+				}
 			}
 		}
 		return user_ids;
